Split feature gates between API server and controller manager
