+++ pkg/controllermanager/features/features.go
@@ -12,35 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package features
 
 import (
-	"flag"
-	"os"
-	"runtime"
-
-	"github.com/gardener/gardener/cmd/gardener-apiserver/app"
-	"github.com/gardener/gardener/pkg/apiserver/features"
-	genericapiserver "k8s.io/apiserver/pkg/server"
-	"k8s.io/apiserver/pkg/util/logs"
+	"github.com/gardener/gardener/pkg/features"
+	utilfeature "k8s.io/apiserver/pkg/util/feature"
 )
 
-func init() {
-	features.RegisterFeatureGates()
-}
-
-func main() {
-	logs.InitLogs()
-	defer logs.FlushLogs()
-
-	if len(os.Getenv("GOMAXPROCS")) == 0 {
-		runtime.GOMAXPROCS(runtime.NumCPU())
+var (
+	// FeatureGate is a shared global FeatureGate for Gardener Controller Manager flags.
+	FeatureGate  = utilfeature.NewFeatureGate()
+	featureGates = map[utilfeature.Feature]utilfeature.FeatureSpec{
+		features.Logging:               {Default: false, PreRelease: utilfeature.Alpha},
+		features.CertificateManagement: {Default: false, PreRelease: utilfeature.Alpha},
 	}
+)
 
-	stopCh := genericapiserver.SetupSignalHandler()
-	command := app.NewCommandStartGardenerAPIServer(os.Stdout, os.Stderr, stopCh)
-	command.Flags().AddGoFlagSet(flag.CommandLine)
-	if err := command.Execute(); err != nil {
-		os.Exit(1)
-	}
+// RegisterFeatureGates registers the feature gates of the Gardener Controller Manager.
+func RegisterFeatureGates() {
+	FeatureGate.Add(featureGates)
 }
