+++ pkg/operation/botanist/controlplane.go
@@ -20,10 +20,12 @@ import (
 	"path/filepath"
 	"strings"
 
+	controllermanagerfeatures "github.com/gardener/gardener/pkg/controllermanager/features"
 	"github.com/gardener/gardener/pkg/features"
 	"github.com/gardener/gardener/pkg/operation/certmanagement"
 	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/utils"
+
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -451,7 +453,7 @@ func (b *Botanist) patchDeploymentCloudProviderChecksums(deploymentName string) 
 
 // DeploySeedLogging will install the Helm release "seed-bootstrap/charts/elastic-kibana-curator" in the Seed clusters.
 func (b *Botanist) DeploySeedLogging() error {
-	if !features.ControllerFeatureGate.Enabled(features.Logging) {
+	if !controllermanagerfeatures.FeatureGate.Enabled(features.Logging) {
 		return common.DeleteLoggingStack(b.K8sSeedClient, b.Shoot.SeedNamespace)
 	}
 
@@ -505,7 +507,7 @@ func (b *Botanist) WakeUpControllers() error {
 
 // DeployCertBroker deploys the Cert-Broker to the Shoot namespace in the Seed.
 func (b *Botanist) DeployCertBroker() error {
-	certManagementEnabled := features.ControllerFeatureGate.Enabled(features.CertificateManagement)
+	certManagementEnabled := controllermanagerfeatures.FeatureGate.Enabled(features.CertificateManagement)
 	if !certManagementEnabled {
 		return b.DeleteCertBroker()
 	}
